test(mysql): cover shared statement classification
